requirements.txt
tellurium>=2.1.0
numpy>=1.15.0 
scipy>=0.19.0  # 0.19.1
matplotlib>=2.0.2
libroadrunner>=1.4.24
antimony>=2.9.4